Separate host function value type ownership
- Move host ABI value type metadata into its own function child module.
- Keep schema assembly and plan-type conversion under explicit owners.
